+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Expose `setSticky`/`setUnsticky` methods in StickyContainer.
+
 ## [2.7.0] - 2019-05-24
 
 Features: